requirements.txt
aioreactive==0.15.0
asyncio==3.4.3
asyncstdlib==3.9.0
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Expression==0.46.0
immutabledict==1.2.0
iniconfig==1.1.1
lark-parser==0.11.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-asyncio==0.14.0
pyzmq==21.0.0
toml==0.10.2
typing-extensions==3.7.4.3
